### 2024-09-25 #90 @Thiggel
[Re] Learning Fair Graph Representations via Automated Data Augmentations
**Original article:**

**Scientific domain:**  Artificial Intelligence
