### 2024-10-11 #90 @Thiggel
Yes it was, the review can be found here: https://openreview.net/forum?id=9OlU865dAF
