### 2025-01-06 #90 @rougier
Oh nice, so we don't need to review again. What is not clear to me is the reason TMLR rejected your paper in the first place. Anyway, based on the reviews and the revisions you made, I think we can publish it provided you use the ReScience template. Can you adapt the paper and possibly put it in a dedicated repo (or directory in your project) since I'll need to fork it for publication.
