+++ rescience_submission/metadata.tex
@@ -1,0 +1,52 @@
+% DO NOT EDIT - automatically generated from metadata.yaml
+
+\def \codeURL{https://zenodo.org/records/13834566}
+\def \codeDOI{10.5281/zenodo.13834566}
+\def \codeSWH{}
+\def \dataURL{https://zenodo.org/records/13837423}
+\def \dataDOI{10.5281/zenodo.13837423}
+\def \editorNAME{Nicolas P. Rougier}
+\def \editorORCID{}
+\def \reviewerINAME{}
+\def \reviewerIORCID{}
+\def \reviewerIINAME{}
+\def \reviewerIIORCID{}
+\def \dateRECEIVED{25 September 2024}
+\def \dateACCEPTED{06 January 2025}
+\def \datePUBLISHED{}
+\def \articleTITLE{[Re] Learning Fair Graph Representations via Automated Data Augmentations}
+\def \articleTYPE{Replication}
+\def \articleDOMAIN{Artificial Intelligence}
+\def \articleBIBLIOGRAPHY{main.bib}
+\def \articleYEAR{2025}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/90}
+\def \articleABSTRACT{Due to the expanding reproducibility crisis in machine learning, there is a pressing need for careful evaluation of research findings.
+We evaluate the reproducibility of the paper \enquote{Learning Fair Graph Representations via Automated Data Augmentations} by \citet{ling2023learning}. Our objective is to assess the authors' three major claims: (1) fair augmentations improve fairness while retaining similar accuracy compared to other fairness methods, (2) augmenting both edges and node features performs better than augmenting only one of the two, and (3) learned augmentations reduce node-wise sensitive homophily and correlation between node features and the sensitive attribute.
+To achieve this, we (a) repeat all experiments from the original paper; (b) implement a new multi-run evaluation protocol with different random seeds; (c) perform extensive ablations of components of Graphair, and (d) investigate the generalizability of the method to other graph neural network (GNN) architectures, and graphs with varying homophily.
+Our results indicate that claims (1), (2), and (3) are only partially reproducible, achieving comparable performance on a maximum of two out of the three original datasets. 
+Notably, the specific dataset(s) where the claims fail to hold varies between experiments.
+Additionally, we found the method generalizes to other GNN architectures, but does not generalize to graphs with varying homophily, failing for unbalanced homophily settings. 
+Finally, throughout our experiments, we uncover a lack of stability in the Graphair framework. }
+\def \replicationCITE{Hongyi Ling, Zhimeng Jiang, Youzhi Luo, Shuiwang Ji, and Na Zou. Learning fair graph representations via automated data augmentations. In The Eleventh International Conference on Learning Representations, 2023}
+\def \replicationBIB{ling2023learning}
+\def \replicationURL{https://openreview.net/pdf?id=1_OGWcP1s9w}
+\def \replicationDOI{No DOI available}
+\def \contactNAME{Max Belitsky}
+\def \contactEMAIL{max.belitsky@student.uva.nl}
+\def \articleKEYWORDS{python, graph, deep learning}
+\def \journalNAME{ReScience C}
+\def \journalVOLUME{4}
+\def \journalISSUE{1}
+\def \articleNUMBER{}
+\def \articleDOI{}
+\def \authorsFULL{Max Belitsky et al.}
+\def \authorsABBRV{M. Belitsky et al.}
+\def \authorsSHORT{Belitsky et al.}
+\title{\articleTITLE}
+\date{}
+\author[1, *]{Max Belitsky}
+\author[1, *]{Filipe Laitenberger}
+\author[1, *]{Denys Sheremet}
[…]
