+++ rescience_submission/content.tex
@@ -1,0 +1,466 @@
+\section{Introduction}
+
+The reproducibility crisis identified in machine learning \citep{kapoor2022leakage} highlights the critical necessity for replication and validation in research. 
+Our investigation focuses on reproducing and extending the work by \citealp{ling2023learning}, which proposed a fair learning method for Graph Neural Networks (GNNs; \citealp{kipf2017semisupervised}).
+
+GNNs have shown considerable performance \citep{gao2019graph, liu2021fast, liu2021dig, yuan2021explainability} across multiple domains \citep{hamaguchi2017knowledge, liu2022spherical, han2022g, hamilton2017inductive}. 
+Despite this success, GNNs have been shown to carry over or amplify bias present in training data \citep{dai2021say, mehrabi2021survey, olteanu2019social}. Specifically, the output of neural networks can be substantially correlated with sensitive features such as ethnicity, nationality, and gender, even though this behavior might not have been intended  \citep{dai2021say}.
+Such biases can be present in graph data, hidden in both the way nodes are connected and their features \citep{mehrabi2021survey, olteanu2019social}. 
+Moreover, removing sensitive features does not always lead to a fairer dataset, as the sensitive information can be retrieved from other features or their combinations \citep{sweeney2002k}.
+
+To tackle this issue, previous approaches for fair graph models have employed various pre-processing techniques (i.e. augmenting the data to remove bias before training the actual network).
+Some augmentation-based methods used handcrafted heuristics for changing the training data \citep{spinelli2021fairdrop, kose2022fair}. Building upon this foundation, \citet{ling2023learning} propose \textit{Graphair}, a framework focused on learning data augmentations that remove bias from a dataset. 
+The authors claim that this method considerably improves on common group fairness metrics across multiple datasets compared to other fair graph training schemes while achieving similar performance in terms of accuracy. We reproduce the experiments by \citet{ling2023learning} to verify their main claims, as well as extend the tests of the method to examine its generalizability and robustness. More specifically, our contributions are summarized below:
+\begin{enumerate}
+    \item We repeat all experiments from the original paper by \citet{ling2023learning}.
+
+    \item We alter the evaluation procedure to retrain the entire method five times with different random seeds, thereby achieving a more robust assessment of the method's stability and performance.
+    
+    \item We assess how well the method generalizes to different architectures.
+    
+    \item We conduct further ablation studies to investigate how the individual components of the Graphair framework contribute to improving fairness.
+    
+    \item We run additional experiments with synthetic datasets to assess the stability and performance of Graphair for datasets with different homophily. 
+    
+    \item We improve the readability and reproducibility of the given code.
+\end{enumerate}
+
+\section{Scope of Reproducibility}
+
+\citet{ling2023learning} propose a technique that automatically learns augmentations that change both the adjacency matrix and the node embeddings of a graph by incorporating adversarial and contrastive learning approaches. We identify the following claims from the paper:
+\begin{itemize}[leftmargin=*]
+    \item[]\textbf{Claim 1:}\claim{claim:1} Learned augmentations \textbf{improve group fairness metrics} while retaining \textbf{similar accuracy} compared to other fair graph-learning methods, on datasets containing sensitive attributes.
[…]
