+++ rescience_submission/article.tex
@@ -1,0 +1,31 @@
+\documentclass{rescience}
+
+\usepackage{hyperref}
+\usepackage{url}
+\usepackage{booktabs}
+\usepackage{multirow}
+\usepackage{subcaption}
+\usepackage{graphicx}
+\usepackage{svg}
+\usepackage{enumitem}
+\usepackage{float}
+\usepackage{makecell}
+\usepackage{csquotes}
+
+\newcounter{claim}
+\newcommand{\claim}[1]{\refstepcounter{claim}\label{#1}}
+
+\begin{document}
+\input{header.tex}
+\input{content.tex}
+
+\hypersetup{linkcolor=black,urlcolor=darkgray}
+\renewcommand\emph[1]{{\bfseries #1}}
+\setlength\bibitemsep{0pt}
+\printbibliography
+
+\input{appendix.tex}
+
+
+
+\end{document}

[…]
