### 2025-03-17 #90 @MaxBelitsky
Hi @rougier, we added a directory with the ReScience template: https://github.com/Thiggel/FACT/tree/main/rescience_submission. Please, let us know if anything else is needed.
