+++ requirements.txt
@@ -5,6 +5,6 @@ scipy==1.11.4
 PyYAML==6.0.1
 matplotlib==3.8.2
 tensorboard==2.14.1
-torch==2.1.2
+torch==2.4.0
 seaborn==0.13.2
 Cython==3.0.8
