### 2025-03-24 #90 @Thiggel
The origin SWHID is:
swh:1:ori:35c0a91c31b06a24561bece63e76ccecc1787bcb

The snapshot SWHID is:
swh:1:snp:894decb239b454eaed7dc5c99d9fe6b5400d7a26

Is this what you mean? or where can I find the swid?
