### 2025-06-23 #90 @Thiggel
Hey, the orcids are as follows:

Max Belitsky: 0009-0000-0396-2916
Nordin Belkacemi: 0009-0005-0288-3024
Filipe Laitenberger: 0009-0003-7836-9828
Denys Sheremet does not have an ORCID id.

Best regards
