### 2025-07-23 #90 @rougier
Thanks. The **sandbox** version of your article is [here](https://sandbox.zenodo.org/record/295008). Can you check everything is ok (especially all links). Once check I'll publish the actual version (with a different DOI)
