### 2025-07-23 #90 @MaxBelitsky
Hi @rougier, we reviewed the sandbox version and found some things that need to be changed:
- The abstract is incorrect – it is from the paper we reproduced, not from our paper
- An indication of equal contribution of all authors is missing
- A comma after the 3rd author's name is missing
- The link to our codebase on page 2 leads to an anonymous repo (we forgot to update it)

Other than that, everything seems to be fine. All the links work.

How can we make the changes? Is there a branch you used that we can create a PR to?
