### 2025-07-23 #90 @rougier
- Weird, I cloned your repo and use your rescience_submission. Did you put your abstract in the matadata.yaml file ? Where is the right abstract then?
- For equal contribution, how does it appear on the paper? Did ou sue special signs?
- I don't quite understand why it is the case.
- Can you give me the right repo then?
