### 2025-07-23 #90 @MaxBelitsky
@rougier

We had problems compiling locally as the script broke, so we used Overleaf and modified the `metadata.tex` manually instead and forgot to modify the YAML file. I've made the following changes to our repo:

- I updated the `metadata.yml` in our repo with the correct abstract
- For equal contribution, I added another affiliation in `metadata.yml`. I hope this would work. When we compiled the document, the `metadata.tex` looked like this:
```
...
\author[1, *]{Max Belitsky}
\author[1, *]{Filipe Laitenberger}
\author[1, *]{Denys Sheremet}
\author[1, *]{Nordin Belkacemi}
\affil[*]{Equal contribution}
\affil[1]{University of Amsterdam, Amsterdam, Netherlands}
```
I added the same symbol in `metadata.yml`. Hopefully, the script would produce the same tex code.

```
authors:
  - name: Max Belitsky
    orcid:
    email: max.belitsky@student.uva.nl
    affiliations: 1,*

  - name: Filipe Laitenberger
    orcid:
    email: filipe.laitenberger@student.uva.nl
    affiliations: 1,*

  - name: Denys Sheremet
    orcid:
    email: denys.sheremet@student.uva.nl
    affiliations: 1,*

  - name: Nordin Belkacemi
    orcid:
    email: nordin.belkacemi@student.uva.nl
    affiliations: 1,*

affiliations:
  - code:    1
    name:    University of Amsterdam
    address: Amsterdam, Netherlands
  - code:    "*"
    name:    Equal contribution
```

- I guess that's not important, we can leave it like this if there is no easy fix
- Fixed the URL in the footnote in our repo
