### 2025-07-23 #90 @rougier
It's online at https://zenodo.org/records/16374814. Congratulations. It will soon appear on ReScience website.
